+++ src/commands/validate/plugin/rules/manifest-action-uuids.ts
