+++ src/commands/validate/plugin/rules/manifest-schema.ts
