+++ src/commands/validate/plugin/validate.ts
@@ -1,11 +1,12 @@
 import type { ValidationResult } from "../result";
 import { validate } from "../validator";
 import { PluginContext } from "./contexts/plugin";
-import { actionUuidIsUniqueAndPrefixed } from "./rules/action-uuid";
-import { manifestExistsAndSchemaIsValid } from "./rules/manifest-exists-and-schema-valid";
+import { actionUuidIsUniqueAndPrefixed } from "./rules/manifest-action-uuids";
 import { manifestFilesExist } from "./rules/manifest-files-exist";
+import { manifestExistsAndSchemaIsValid } from "./rules/manifest-schema";
 import { manifestUrlsExist } from "./rules/manifest-urls-exist";
 import { pathIsDirectoryAndUuid } from "./rules/path-input";
+
 /**
  * Validates the Stream Deck plugin as the specified {@link path}.
  * @param path Path to the plugin.
